refactor(db): centralize raw-SQL into src/db/queries and harden rate-limit
The trigger was a production crash on the v1 API and Plaud sync: every
request that hit the rate-limit bucket upsert returned 500 with
postgres-js's ERR_INVALID_ARG_TYPE ("Received an instance of Date").
Cause: raw `sql\`case when reset_at <= ${now}\`` interpolated a Date
into Drizzle's tagged template, which bypasses Drizzle's column
encoders -- the driver got a Date object and threw.

While fixing it, audited every raw-SQL site in the codebase and pulled
them into a new `src/db/queries/` layer so the surface is reviewable
in one place.

Bug fixes
- Cast Date params to `${iso}::timestamp` in the rate-limit bucket
  upsert (`src/db/queries/rate-limit.ts`). Matches the column type
  (`timestamp`, TZ-naive); `::timestamptz` would skew comparisons
  under non-UTC server timezones.
- Same cast applied to six implicit text->timestamp comparisons in
  the admin analytics queries (`listUsers`, `topServerTranscriptionUsers`,
  `syncHealth`, `pricingSnapshot`). They worked via Postgres's implicit
  coercion but could defeat index usage on `created_at` / `last_sync`.

Safety
- Rate limiter now fails OPEN when the bucket store is unreachable:
  catches the DB error, logs a warning, returns `allowed: true` with
  a synthetic full bucket. A broken safety net should not take down
  the API; upstream (Cloudflare / ALB) still rate-limits at the edge
  and `/api/v1/*` has per-token auth as a second gate.
- Escape `%` and `_` in admin user-search ILIKE pattern with
  `escape '\\'` -- closes a low-severity wildcard-injection footgun.

Refactor
- New `src/db/queries/` layer (5 files):
  - `admin.ts` (moved verbatim from `src/lib/admin/queries.ts`)
  - `admin-pricing-snapshot-csv.ts` (extracted from CSV route)
  - `webhook-deliveries.ts` (claim/release/reload from worker)
  - `rate-limit.ts` (atomic bucket upsert)
  - `plaud-locks.ts` (advisory lock helper)
- Inline `sql\`waveform_peaks is null\`` in peaks route replaced with
  Drizzle's `isNull()` (already imported).
- AGENTS.md documents the new policy: raw-SQL queries
  (`db.execute(sql\`...\`)` and full statements) MUST live in
  `src/db/queries/`. Expression fragments embedded in Drizzle's
  fluent builder (`set: { col: sql\`col + 1\` }`, `where(sql\`a @> b\`)`)
  stay inline. Added a `sql\`\`` security checklist covering param
  binding, Date casts, ILIKE escaping, and identifier allowlists.

Tests
- New regression test (`src/tests/rate-limit-bucket.test.ts`):
  walks the `onConflictDoUpdate.set` SQL fragments and fails if any
  `Date` instance is embedded -- pins the original crash in place.
  Second case asserts fail-open behavior under a thrown DB error.
- Updated PII guard test to read the moved file path.
- `pnpm type-check` and `pnpm test` both green (371 tests pass,
  5 opt-in integration tests skipped).
