+++ src/lib/webhooks/worker.ts
@@ -2,8 +2,13 @@ import type { IncomingMessage, RequestOptions } from "node:http";
 import { request as httpRequest } from "node:http";
 import { request as httpsRequest } from "node:https";
 import type { LookupFunction } from "node:net";
-import { and, eq, inArray, lte, or, type SQL, sql } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { db } from "@/db";
+import {
+    claimDueWebhookDeliveries,
+    releaseClaimedDelivery,
+    reloadClaimedDeliveryForSend,
+} from "@/db/queries/webhook-deliveries";
 import { webhookDeliveries, webhookEndpoints } from "@/db/schema";
 import {
     createOutboundWebhookPayload,
@@ -28,10 +33,7 @@ import {
 } from "@/lib/webhooks/url";
 
 const TICK_MS = 30_000;
-const DELIVERY_LIMIT = 50;
-const PER_USER_DELIVERY_LIMIT = 10;
 const TIMEOUT_MS = 10_000;
-const PROCESSING_LEASE_MS = 15 * 60_000;
 const BACKOFF_MS = [30_000, 120_000, 600_000, 3_600_000, 21_600_000] as const;
 
 let started = false;
@@ -46,38 +48,10 @@ type WebhookDeliveryResult = {
     permanentFailure?: boolean;
 };
 
-type ClaimedDelivery = {
-    delivery: typeof webhookDeliveries.$inferSelect;
-    endpoint: typeof webhookEndpoints.$inferSelect;
-};
-
-type CandidateDeliveryRow = {
-    id: string;
-};
-
-type QueryResultRows<T> = T[] | { rows: T[] };
-
 export function getWebhookBackoffMs(attemptNumber: number): number {
     return BACKOFF_MS[
         Math.min(Math.max(attemptNumber, 1), BACKOFF_MS.length) - 1
     ];
-}
-
-function rowsFromQueryResult<T>(result: QueryResultRows<T>): T[] {
-    return Array.isArray(result) ? result : result.rows;
-}
-
-function dueDeliveryPredicate(now: Date): SQL {
-    return or(
-        and(
-            eq(webhookDeliveries.status, "pending"),
-            lte(webhookDeliveries.nextAttemptAt, now),
-        ),
-        and(
-            eq(webhookDeliveries.status, "processing"),
-            lte(webhookDeliveries.nextAttemptAt, now),
-        ),
-    ) as SQL;
 }
 
 async function readResponseBody(response: IncomingMessage): Promise<string> {
@@ -361,150 +335,6 @@ async function markDeliveryAttempt(
                 ),
             );
     });
-}
-
-async function claimDueWebhookDeliveries(): Promise<ClaimedDelivery[]> {
-    const now = new Date();
-    // postgres-js binds Date → timestamp parameters reliably when the
-    // Date sits behind a Drizzle column predicate, but raw sql`...` with
-    // Date placeholders crashes under Bun/Next 16 with
-    // ERR_INVALID_ARG_TYPE ("Received an instance of Date"). Cast the
-    // ISO string explicitly to `timestamp` to match the column type
-    // (`webhookDeliveries.nextAttemptAt` is `timestamp`, not
-    // `timestamptz`); using `::timestamptz` would force a timezone
-    // conversion on every comparison and skew the due window in any
-    // non-UTC server timezone.
-    const nowParam = sql`${now.toISOString()}::timestamp`;
-
-    const candidateResult = await db.execute(sql`
-        select id
-        from (
-            select
-                ${webhookDeliveries.id} as id,
-                ${webhookDeliveries.nextAttemptAt} as next_attempt_at,
-                row_number() over (
-                    partition by ${webhookDeliveries.userId}
-                    order by ${webhookDeliveries.nextAttemptAt} asc, ${webhookDeliveries.id} asc
-                ) as user_rank
-            from ${webhookDeliveries}
-            inner join ${webhookEndpoints}
-                on ${webhookEndpoints.id} = ${webhookDeliveries.endpointId}
-            where (
-                (${webhookDeliveries.status} = 'pending' and ${webhookDeliveries.nextAttemptAt} <= ${nowParam})
-                or (${webhookDeliveries.status} = 'processing' and ${webhookDeliveries.nextAttemptAt} <= ${nowParam})
-            )
-            and ${webhookEndpoints.enabled} = true
-        ) ranked_deliveries
-        where user_rank <= ${PER_USER_DELIVERY_LIMIT}
-        order by next_attempt_at asc, id asc
-        limit ${DELIVERY_LIMIT}
-    `);
-
-    const candidateRows = rowsFromQueryResult(
-        candidateResult as unknown as QueryResultRows<CandidateDeliveryRow>,
-    );
-
-    if (candidateRows.length === 0) return [];
-
-    const ids = candidateRows.map((row) => row.id);
-    const claimExpiresAt = new Date(now.getTime() + PROCESSING_LEASE_MS);
-    const claimedRows = await db
-        .update(webhookDeliveries)
-        .set({
-            status: "processing",
-            nextAttemptAt: claimExpiresAt,
-            updatedAt: now,
-        })
-        .where(
-            and(
-                inArray(webhookDeliveries.id, ids),
-                dueDeliveryPredicate(now),
-                sql`exists (
-                    select 1
-                    from ${webhookEndpoints}
-                    where ${webhookEndpoints.id} = ${webhookDeliveries.endpointId}
-                    and ${webhookEndpoints.enabled} = true
-                )`,
-            ),
-        )
-        .returning({ id: webhookDeliveries.id });
-
-    const claimedIds = new Set(claimedRows.map((row) => row.id));
-    if (claimedIds.size === 0) return [];
-
-    const rows = await db
-        .select({
-            delivery: webhookDeliveries,
-            endpoint: webhookEndpoints,
-        })
-        .from(webhookDeliveries)
-        .innerJoin(
-            webhookEndpoints,
-            eq(webhookEndpoints.id, webhookDeliveries.endpointId),
-        )
-        .where(
-            and(
-                inArray(webhookDeliveries.id, Array.from(claimedIds)),
-                eq(webhookEndpoints.enabled, true),
-            ),
-        );
-
-    const order = new Map(ids.map((id, index) => [id, index]));
-    return rows.sort((a, b) => {
-        return (
-            (order.get(a.delivery.id) ?? Number.MAX_SAFE_INTEGER) -
-            (order.get(b.delivery.id) ?? Number.MAX_SAFE_INTEGER)
-        );
-    });
-}
-
-async function reloadClaimedDeliveryForSend(
-    claimed: ClaimedDelivery,
-): Promise<ClaimedDelivery | null> {
-    const [row] = await db
-        .select({
-            delivery: webhookDeliveries,
-            endpoint: webhookEndpoints,
-        })
-        .from(webhookDeliveries)
-        .innerJoin(
-            webhookEndpoints,
-            eq(webhookEndpoints.id, webhookDeliveries.endpointId),
-        )
-        .where(
-            and(
-                eq(webhookDeliveries.id, claimed.delivery.id),
-                eq(webhookDeliveries.userId, claimed.delivery.userId),
-                eq(webhookDeliveries.endpointId, claimed.endpoint.id),
-                eq(webhookDeliveries.status, "processing"),
-                eq(webhookEndpoints.id, claimed.endpoint.id),
-                eq(webhookEndpoints.userId, claimed.endpoint.userId),
-                eq(webhookEndpoints.enabled, true),
-            ),
-        )
-        .limit(1);
-
-    return row ?? null;
-}
-
-async function releaseClaimedDelivery(
-    delivery: typeof webhookDeliveries.$inferSelect,
-): Promise<void> {
-    const now = new Date();
-    await db
-        .update(webhookDeliveries)
-        .set({
-            status: "pending",
-            nextAttemptAt: now,
-            updatedAt: now,
-        })
-        .where(
-            and(
-                eq(webhookDeliveries.id, delivery.id),
-                eq(webhookDeliveries.userId, delivery.userId),
-                eq(webhookDeliveries.status, "processing"),
-            ),
-        );
 }
 
 export async function deliverDueWebhooks(): Promise<void> {
